Fixed issues with long/double math in Expiration timeout

+++ src/main/kotlin/edu/csh/chase/sprint/BackoffTimeout.kt
@@ -58,7 +58,7 @@ abstract class BackoffTimeout(val maxAttempts: Int?) {
             unit.toMillis(start), power, unit.toMillis(maxTimeout), maxAttempts
         )
 
-        override fun getDelay(attempt: Int): Long = Math.min(start * Math.pow(power.toDouble(), attempt.toDouble()).toLong(), maxTimeout)
+        override fun getDelay(attempt: Int): Long = Math.min(start.toDouble() * Math.pow(power.toDouble(), attempt.toDouble()), maxTimeout.toDouble()).toLong()
 
     }
